requirements.txt
PyJWT == 0.2.1
oauthlib == 0.6.1
requests == 2.2.1
requests-oauthlib == 0.4.0
six == 1.5.2
